Refactor internals
Instead of heavily nested procs, this commit sort of flattens out the
structure. For the following structure:

    {
      data: [
        {
          id: "1",
          type: "articles",
          attributes: {
            title: "hello world",
            published_at: "2018-07-01T09:00:00Z",
          },
          relationships: {
            author: { data: { id: "1", type: "people" } },
            comments: { data: [{ id: "1", type: "comment" }] },
          },
        },
        ...
      ],
      included: {
        authors: {
          data: [
            {
              id: "1",
              type: "people",
              attributes: {
                name: "Jamie",
                email: "jamie@example.org",
              },
            },
          ],
        },
        comments: {
          data: [
            {
              id: "1",
              type: "comments",
              attributes: {
                body: "hi",
                created_at: "2018-07-01T10:00:00Z",
              },
              relationships: {
                author: { data: { id: "2", type: "people" } },
              },
            },
            ...
          ],
        },
      }
    }

We can think of it as the hashes with a `data` key being the result of a
`DataSerializer`, the {id:, type:} subhash being a `MetadataSerializer`,
and a `ModelSerializer` being a superset of that (including `attributes`
and `relationships`, which are `AttributesSerializer` and
`RelationshipsSerializer`, respectively). Then, the `included` hash is
the same thing applied to other models. This radically simplifies
serialization.

Everything is still type-checked, but we're probably doing even more type
checking now than we need.

This commit also eliminates caching for now because it appears to slow
things down with this implementation. I haven't figured out why yet,
though.